Response request dto

+++ src/main/java/com/example/spring/demo/projectmanagement/entities/Project.java
@@ -1,7 +1,8 @@
 package com.example.spring.demo.projectmanagement.entities;
 
-import com.example.spring.demo.projectmanagement.entities.Employee;
-import com.fasterxml.jackson.annotation.*;
+import com.fasterxml.jackson.annotation.JsonIdentityInfo;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import jakarta.persistence.*;
 
 import java.util.ArrayList;
@@ -13,7 +14,7 @@ import java.util.List;
 public class Project {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @Column(unique = true)
     private String name;
@@ -36,11 +37,11 @@ public class Project {
         this.employeeList = employeeList;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
